pyproject.toml
[tool.poetry]
name = "hometube"
version = "0.1.0"
description = ""
authors = ["Huy Nguyen <giahuy2201@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
yt-dlp = "^2023.12.30"
SQLAlchemy = "^2.0.25"
pillow = "^10.2.0"
pydantic-settings = "^2.1.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.1"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
